params: remove useless casts

+++ libnopegl/src/params.c
@@ -858,14 +858,14 @@ int ngli_params_set_defaults(uint8_t *base_ptr, const struct node_param *params)
         int ret = 0;
         switch (par->type) {
             case NGLI_PARAM_TYPE_SELECT: {
-                const int v = (int)par->def_value.i32;
+                const int v = par->def_value.i32;
                 const char *s = ngli_params_get_select_str(par->choices->consts, v);
                 ngli_assert(s);
                 ret = ngli_params_set_select(dstp, par, s);
                 break;
             }
             case NGLI_PARAM_TYPE_FLAGS: {
-                const int v = (int)par->def_value.i32;
+                const int v = par->def_value.i32;
                 char *s = ngli_params_get_flags_str(par->choices->consts, v);
                 if (!s)
                     return NGL_ERROR_INVALID_ARG;
